fixed some bugs in the expression parser, and plugged all the memory leaks. The expression parser now will not leak, as long as there are no syntax errors. Sorry, but freeing tokens on error is a pretty complex affair. Maybe some future rev. Usually the amount leaked on syntax errors is fairly small.
Added checking, fixed the makefile to make the check_expr.c
program. Added an extra entry to make "testexpr2s", a new
exec that tests all the expressions in a file, where the filename
is the only argument.

The expr2.testinput is a set of expressions to provide some regression 
testing for the expr2 stuff.

Also noted that ast_expr2.fl was not returning the =~ operator,
or the ~ operator, and had an entry in the TOKEN pattern that would 
glom a % to the nearest token. I considered all these as bugs!