Add proper codec support to chan_skinny.  Works with at least ulaw, alaw, and g729a.
This is technically a "new feature", but there are justifications for it.

I found a bug with the recent rtp packetization changes, which caused the media setup to
fail under certain circumstances, particularly when using allow=all, or having no allow=
statements (globally or on the device).

I could have either removed the rtp packetization features, or I could add proper codec
support (which, without, I think most people would consider to be a bug anyways).