Merged revisions 43699 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r43699 | russell | 2006-09-26 16:23:15 -0400 (Tue, 26 Sep 2006) | 6 lines

When parsing the sections of voicemail.conf that contain mailbox definitions,
don't introduce a length limit on the definition by using a 256 byte temporary
storage buffer.  Instead, make the temporary buffer just as big as it needs
to be to hold the entire mailbox definition.
(fixes BE-68)

........