Merged revisions 43739 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r43739 | murf | 2006-09-26 20:32:47 -0600 (Tue, 26 Sep 2006) | 1 line

This change to extensions.ael was to fix bug 8031; the install scripts are causing it to be copied to /etc/asterisk/extensions.ael, and because it is a fairly direct conversion of the original extensions.conf, the macro and context names clash with the existing extensions.conf. So, I put an ael- in front of all macros and contexts, and checked every goto and macro call. Also, this file compiles under aelparse.
........