Trying to add some sense of direction in a transaction.
If we have an OUTBOUND call, we still may get a RE-INVITE which
creates an inbound transaction.

On an INBOUND call, we might send a RE-INVITE, which creates
an outbound transaction.

This outbound/inbound stuff confuses the CANCEL and the ACK
and mixes some TO/FROM headers and tags.

This patch is *not* tested yet.