Merged revisions 46845 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r46845 | russell | 2006-11-01 17:32:12 -0500 (Wed, 01 Nov 2006) | 5 lines

Add a check in the configure script to determine whether ld is GNU ld or not.
This is needed because module embedding only works for gnu ld.  GNU ld is now
listed as a dependency for all of the module embedding options in menuselect.
(issue #8143)

........