start integrating the simplifications proposed in bug 0005860, as usual a bit at a time to ease locating new bugs or fixes worth merging into other branches.
In this commit, introduce a macro, S_REPLACE, that replaces
a string possibly freeing the previous value.
In one of these places (see the comment marked XXX) the previous
code might leak memory - if so, this ought to be merged in 1.4

The macro might be worth putting in one of the global headers
(e.g. include/asterisk/strings.h) as the construct is used
in a million places in the asterisk code.