put the body of "sip show channel" and "sip show history" in separate functions, to make it easy to move the dialoglist to use containers, either with iterators or callbacks.
Note that currently (and in all versions so far)
these two functions, which are expensive and can take a long
time to do the I/O, run with the list locked.
This means that during their execution, incoming packets cannot
be processed (find_call() blocks on the list).

Moving the implementation to use iterators will remove this
problem (this commit is a step in that direction, but we are not there yet).