prepare the conversion of the dialoglist to astobj2. All the new code is commented out so still works as before. If you uncomment the #define USE_AO2 line, the code still compiles but doesn't run because sip_pvt are not yet allocated correctly.
The goal here is to see how the astobj2 API match with
common usage patterns. Ideally, you should be able to
switch from hand-written lists to astobj2 with minimal
differences in the sources, basically just the for()
loop heading, and the constructor/destructors.

And, if the API is correct, the locking requirements with
astobj2 should be a lot lower because many cases hwere we
currently lock things are handled by keeping a refcount.