rename the "owner" field in a sip_pkt to "pvt", as it is too confusing otherwise ("owner" is used to indicate a channel in a nearby struct). Most of the usages are in a single function.
Trunk candidate, definitely.

On passing note that __sip_ack(), when called by handle_response()
with resp == 491, does not free the packet, but because
the packet is not linked in the list anymore, this possibly
causes a memory leak. If not, it is a very obscure thing.

Need to consider this as a possible bug in trunk.