Merged revisions 47391 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.4
........
r47391 | russell | 2006-11-09 16:26:27 -0500 (Thu, 09 Nov 2006) | 7 lines

Work around an issue that caused menuselect to display a bogus description for
app_voicemail and chan_zap.  These modules use some preprocessor directives to
determine what it will report to Asterisk as its description.  However, the way
we extract this information from the source files for menuselect is not smart
enough to figure this out.
(issue #8326, #8328)

........