introduce a bit of regexp support in the internal CLI api. Now you can specify a cli command as 	"console autoanswer [on|off]"
which means the on|off argument is optional, or

	"console {mute|unmute}"

which means the mute|unmute argument is mandatory.

The blocks in [] or {} do not necessarily need to be at the
end of the string.

Completions for the variant parts are generated automatically.
This should significantly simplify the implementation of
the various handlers.