replace a sizeof(struct sip_request) with sizeof(*req);
introduce a new struct to store outgoing packets.
First, because the internal format for incoming and outgoing packets
is different, it is very confusing (and dangerous) to store
them in the same struct, as it prevents the compiler from
helping us.
Second, by using a different structure we can optimize them
for the two situations.