Merged revisions 51085,51145,51158,51161,51194,51197 via svnmerge from https://origsvn.digium.com/svn/asterisk/branches/1.2
........
r51085 | file | 2007-01-15 22:53:31 -0700 (Mon, 15 Jan 2007) | 2 lines

Add none as a valid callgroup/pickupgroup option. I consider it a bug that it would inherit it all the way down and not have any way to reset it to nothing - so that's why it is in 1.2. (issue #8296 reported by gkloepfer)

........
r51145 | file | 2007-01-16 10:36:50 -0700 (Tue, 16 Jan 2007) | 2 lines

Return previous behavior. ParkedCalls will be able to do DTMF based transfers again. trunk however will get an option to allow this to be set on/off. (issue #8804 reported by nortex)

........
r51158 | tilghman | 2007-01-16 14:26:06 -0700 (Tue, 16 Jan 2007) | 2 lines

Postgres driver doesn't like a NULL pointer when retrieving the length (Bug 8513)

........
r51161 | tilghman | 2007-01-16 14:50:04 -0700 (Tue, 16 Jan 2007) | 2 lines

Add documentation walkthrough on getting Postgres to work with voicemail (from Issue 8513)

........
r51194 | tilghman | 2007-01-17 13:52:21 -0700 (Wed, 17 Jan 2007) | 4 lines

When ast_strip_quoted was called with a zero-length string, it would treat a
NULL as if it were the quoting character (and would thus return the string
in memory immediately following the passed-in string).

........
r51197 | russell | 2007-01-17 14:17:21 -0700 (Wed, 17 Jan 2007) | 3 lines

Move the check for a failure of ast_channel_alloc() to before locking the
pvt structure again.  Otherwise, on a failure, this will cause a deadlock.

........