Significantly restructure the SLA implementation ... again.
This new architecture properly handles multiple trunks ringing at the same
time.  Are you wondering why this was a complicated thing to handle?  Well,
let me illustrate a simple, yet complicated scenario.

A call comes in on line2.  This causes outbound calls to station1 and
station2.  However, before either station answers, line1 starts ringing.
Now, when one of the phones answers, it should pick up line1, not line2.
Furthermore, the other phone needs to keep ringing.  And thorough all of this,
all of the blinky lights have to be correct.

Anyway, it's all happy now.