Untested additional feature (idea from Frank Sautter, implementation by moi)
New dial flag to always enable the Answered Elsewhere header in SIP to avoid "missed calls"
on the display. This is useful for dialplans when dialling continues in other steps in a followme-
like fashion.

Maybe this could be implemented in Queue and app_followme() too, but that's something I
happily leave in the hands of BJ... :-)