Merged revisions 75163-75164,75183 via svnmerge from https://origsvn.digium.com/svn/asterisk/trunk
........
r75163 | russell | 2007-07-15 20:44:49 -0600 (Sun, 15 Jul 2007) | 4 lines

Cast the 2nd argument to iconv() to a void *, as some systems define it as a
(const char *), while others define it as (char *).  This is done to suppress
compiler warnings about it.

........
r75164 | russell | 2007-07-15 20:51:56 -0600 (Sun, 15 Jul 2007) | 5 lines

Merge a bunch of doxygen updates to header files.  This includes changes to
use the \retval tag for documenting return values, fixing various warnings
when generating the documentation, and various other things.
(closes issue #10203, snuffy)

........
r75183 | file | 2007-07-16 07:35:20 -0600 (Mon, 16 Jul 2007) | 2 lines

It is no longer required for each module that deals with a channel to call ast_module_user_hangup_all in it's unload function. The loader will automatically perform this action for it.

........