Allow ao2_find to take advantage of the fact that the peers are stored in a hash table.  To do this, I made a slight change to astobj2 described below.
Previously, OBJ_POINTER implied both that the argument provided to ao2_callback
was of the same type of the objects in the container, but *also* implied that
this was the object that was being looked for in the container.  I made a slight
change to ao2_callback such that only the first part was true.  This made it so
I could pass in a temporary object that could be sent to the hash function for
getting the right hash for where to start looking, but let it still use my own
custom object compare function.

I also updated the documentation for the OBJ flags to reflect my changes as well
as an incorrect description for OBJ_NODATA.  OBJ_NODATA means *don't* return the
object (and consequently, increase its reference count) on a match.